+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Force stopping the MoTAN devices when homing.
 * Force telescope homing if the clean-up recipe is run more than once during troubleshooting.
 * Stop observing if the clean-up recipe is run more than three times without success.
+* The ping task now check if the Overwatcher main task has stalled and restarts it if necessary.
 
 ### 🔧 Fixed